+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.yasu_k.saezuri">
+    package="com.yasunari_k.saezuri">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -10,9 +10,9 @@
         android:required="false" />
 
     <application
-        android:name=".MyApplication"
+        android:name="com.yasunari_k.saezuri.MyApplication"
         android:allowBackup="true"
-        android:icon="@drawable/bird_tweeting"
+        android:icon="@drawable/bird_tweeting_2"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
@@ -22,7 +22,7 @@
 
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="com.yasu_k.saezuri.fileprovider"
+            android:authorities="com.yasunari_k.saezuri.fileprovider"
             android:exported="false"
             android:grantUriPermissions="true">
             <meta-data
@@ -30,7 +30,7 @@
                 android:resource="@xml/provider_path"/>
         </provider>
 
-        <activity android:name="com.yasu_k.saezuri.Tweet"
+        <activity android:name="com.yasunari_k.saezuri.Tweet"
             android:exported="true"
             android:hardwareAccelerated="true">
             <intent-filter>
