--- app/src/main/java/com/yasu_k/saezuri/di/SharedPreferencesModule.kt
@@ -1,17 +1,0 @@
-package com.yasu_k.saezuri.di
-
-import android.content.Context
-import android.content.SharedPreferences
-import androidx.appcompat.app.AppCompatActivity
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import javax.inject.Inject
-
-@Module
-class SharedPreferencesModule {
-    @Provides
-    fun provideSharedPreferences(context: Context) : SharedPreferences {
-        return context.getSharedPreferences("twitterToken", AppCompatActivity.MODE_PRIVATE)
-    }
-}
