--- app/src/main/java/com/yasu_k/saezuri/ui/DialogViewModel.kt
@@ -1,7 +1,0 @@
-package com.yasu_k.saezuri.ui
-
-import androidx.lifecycle.ViewModel
-
-class DialogViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
