+++ app/src/main/java/com/yasunari_k/saezuri/ui/DialogViewModel.kt
@@ -1,0 +1,7 @@
+package com.yasunari_k.saezuri.ui
+
+import androidx.lifecycle.ViewModel
+
+class DialogViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
