+++ app/src/main/res/layout/fragment_dialog.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MediaDialogFragment">
+    tools:context="com.yasunari_k.saezuri.ui.MediaDialogFragment">
 
     <TextView
         android:layout_width="match_parent"
