+++ app/src/main/res/layout/tweet.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Tweet">
+    tools:context="com.yasunari_k.saezuri.Tweet">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
