+++ release/output-metadata.json
@@ -4,7 +4,7 @@
     "type": "APK",
     "kind": "Directory"
   },
-  "applicationId": "com.yasu_k.saezuri",
+  "applicationId": "com.yasunari_k.saezuri",
   "variantName": "release",
   "elements": [
     {
